training/requirements.txt
tensorflow-gpu
keras
numpy
Pillow
imutils